changed vars to properties (big bug fixed)
the bug in the previous commit has not been fixed by this, but I found
the root cause while continue working on it.
It was due to Xcode 12 Command Line Tools breaking tweaks building in
arm64e. Using CLT of Xcode 11 solves the problem (see my tweets about
that for more info)